♻️ refactor : #36 Card 리팩토링

+++ src/main/java/com/sparta/refrigerator/card/service/CardService.java
@@ -28,9 +28,9 @@ public class CardService {
     private final CardRepository cardRepository;
     private final ColumnService columnService;
     private final InvitationRepository invitationRepository;
-    private final ColumnRepository columnRepository;
     private final UserService userService;
     private final BoardService boardService;
+    private final ColumnRepository columnRepository;
 
 
     @Transactional
@@ -38,6 +38,15 @@ public class CardService {
         User user) {
         Columns checkColumn = columnService.findById(columnId);
         Board checkBoard = boardService.findById(boardId);
+
+        if (requestDto.getTitle().isEmpty()) {
+            throw new DataNotFoundException("제목을 입력해주기 바랍니다.");
+        }
+
+        if (requestDto.getContent().isEmpty()) {
+            throw new DataNotFoundException("내용을 입력해주시기 바랍니다");
+        }
+
         boolean isBoardUser = invitationRepository.existsByBoardAndUser(checkColumn.getBoard(),
             user);
 
@@ -54,10 +63,19 @@ public class CardService {
 
     @Transactional
     public CardResponseDto updateCard(Long boardId, Long columnId, Long cardId,
-        CardRequestDto requestDto,
-        User user) {
+        CardRequestDto requestDto, User user) {
         Board checkBoard = boardService.findById(boardId);
+
+        if (checkBoard == null) {
+            throw new DataNotFoundException("존재하지 않는 보드입니다");
+        }
+
         Columns checkColumn = columnService.findById(columnId);
+
+        if (checkColumn == null) {
+            throw new DataNotFoundException("존재하지 않는 컬럼입니다");
+        }
+
         Card card = findCard(cardId);
 
         boolean isBoardUser = invitationRepository.existsByBoardAndUser(checkColumn.getBoard(),
@@ -77,23 +95,45 @@ public class CardService {
     @Transactional
     public void deleteCard(Long boardId, Long columnId, Long cardId, User user) {
         Board checkBoard = boardService.findById(boardId);
+
+        if (checkBoard == null) {
+            throw new DataNotFoundException("존재하지 않는 보드입니다");
+        }
+        Columns checkColumn = columnService.findById(columnId);
+
+        if (checkColumn == null) {
+            throw new DataNotFoundException("존재하지 않는 컬럼입니다");
+        }
+
         Card card = cardRepository.findById(cardId)
             .orElseThrow(() -> new CardNotFoundException("해당 카드는 삭제되었거나 존재하지 않습니다."));
         if (!card.getUser().getId().equals(user.getId())) {
-            throw new UserMisMatchException("카드는 생성한 작업자만 수정가능합니다");
+            throw new UserMisMatchException("카드는 생성한 작업자만 삭제가능합니다");
         }
         cardRepository.delete(card);
     }
 
     @Transactional(readOnly = true)
-    public CardResponseDto getCard(Long cardId, User user, Long columnId) {
+    public CardResponseDto getCard(Long cardId, User user, Long columnId, Long boardId) {
+        Board checkBoard = boardService.findById(boardId);
+
+        if (checkBoard == null) {
+            throw new DataNotFoundException("존재하지 않는 보드입니다");
+        }
+
         Columns checkColumn = columnService.findById(columnId);
+
+        if (checkColumn == null) {
+            throw new DataNotFoundException("존재하지 않는 칼럼입니다");
+        }
+
         boolean isBoardUser = invitationRepository.existsByBoardAndUser(checkColumn.getBoard(),
             user);
 
         if (!isBoardUser) {
             throw new DataNotFoundException("보드에 초대된 사용자가 아닙니다");
         }
+
         userService.findById(user.getId());
         return cardRepository.findById(cardId)
             .map(CardResponseDto::new)
@@ -101,19 +141,9 @@ public class CardService {
     }
 
     @Transactional(readOnly = true)
-    public List<CardResponseDto> getCards() {
+    public List<CardResponseDto> getCards(User user) {
+        userService.findById(user.getId());
         List<Card> cards = cardRepository.findAll();
-        return cards
-            .stream()
-            .map(CardResponseDto::new)
-            .collect(Collectors.toList());
-    }
-
-    @Transactional(readOnly = true)
-    public List<CardResponseDto> getCollaboratorCard(Long userId, Long boardId) {
-        Board checkBoard = boardService.findById(boardId);
-        User user = userService.findById(userId);
-        List<Card> cards = cardRepository.findByUser(user);
 
         return cards
             .stream()
@@ -122,10 +152,61 @@ public class CardService {
     }
 
     @Transactional(readOnly = true)
-    public List<CardResponseDto> getColumnNameCard(Long columnId, Long boardId) {
+    public List<CardResponseDto> getBoardCards(User user, Long boardId) {
         Board checkBoard = boardService.findById(boardId);
-        Columns columns = columnService.findById(columnId);
+        userService.findById(user.getId());
+
+        if (checkBoard == null) {
+            throw new DataNotFoundException("보드가 존재하지 않습니다");
+        }
+
+        List<Card> cards = cardRepository.findAllByBoardId(boardId);
+
+        if (cards == null) {
+            throw new DataNotFoundException("해당 보드에는 카드가 존재하지않습니다");
+        }
+
+        return cards
+            .stream()
+            .map(CardResponseDto::new)
+            .collect(Collectors.toList());
+    }
+
+    @Transactional(readOnly = true)
+    public List<CardResponseDto> getAssigneeCard(String collaborator, Long boardId, User user) {
+        Board checkBoard = boardService.findById(boardId);
+        userService.findById(user.getId());
+
+        if (checkBoard == null) {
+            throw new DataNotFoundException("보드가 존재하지 않습니다");
+        }
+
+        List<Card> cards = cardRepository.findAllByCollaboratorAndBoardId(collaborator, boardId);
+
+        if (cards.isEmpty()) {
+            throw new DataNotFoundException("해당 보드에서 작업자가 생성한 카드를 찾을 수 없습니다.");
+        }
+
+        return cards
+            .stream()
+            .map(CardResponseDto::new)
+            .collect(Collectors.toList());
+    }
+
+    @Transactional(readOnly = true)
+    public List<CardResponseDto> getColumnNameCard(Long boardId, String columnName, User user) {
+        Board checkBoard = boardService.findById(boardId);
+        userService.findById(user.getId());
+
+        if (checkBoard == null) {
+            throw new DataNotFoundException("보드가 존재하지 않습니다");
+        }
+
+        Columns columns = columnRepository.findByColumnName(columnName)
+            .orElseThrow(() -> new IllegalArgumentException("컬럼이 없습니다"));
+
         List<Card> cards = cardRepository.findByColumns(columns);
+
         return cards
             .stream()
             .map(CardResponseDto::new)
@@ -137,5 +218,4 @@ public class CardService {
             () -> new DataNotFoundException("해당 카드는 삭제되었거나 존재하지 않습니다.")
         );
     }
-
 }
